Remove base unicode TH
The runUnicodeFunctionBase function -- intended to run unicode
functions with base's version or die with a type error -- was
problematic from a forwards-compatibility perspective.

First, it requires the baseUnicode splice, which will fail if base's
version is not supported. We guarded against this with cpp, but
when a new base comes out with a new unicode, the library will fail
to compile. This is unfortunate, because it means users of the
perfectly fine runUnicodeFunction and runUnicodeFunctionVersion will
also be impacted, even though those functions still work.

We could guard against it with cpp i.e. only compile the splice for
"known good" versions. But then the runUnicodeFunctionBase function
itself has dubious utility, because you cannot use it when a new
unicode version comes out, until it is supported.

We therefore remove runUnicodeFunctionBase, expecting
runUnicodeFunction to be the default choice.

If someone really wants the "base or die" functionality, they can
write it themselves easily using the provided utilities.