Remove 'base' unicode version functions
Removes the functionality that retrieves the unicode version 'base'.
Unfortunately, the combination of:

1. unicode-grapheme being compatible with multiple major versions of
   base.

2. unicode-grapheme can use base's unicode version.

violates PVP. More precisely, the behavior of unicode-grapheme can
depend on the major version of base. This is by design, but what went
unnoticed is that this violates PVP, as a package's API (which includes
semantic behavior) can only depend on _its_ version number:

  "It’s the responsibility of the provider to guarantee that the exposed
  API is a function of the declared version, and in particular does not
  depend on the versions of its transitive dependencies, as this defy
  the purpose of the PVP."

https://pvp.haskell.org/faq/#what-implications-does-the-pvp-have-when-re-exporting-api-elements

The API we want to provide i.e. "this function's behavior is relative to
base's unicode version" unfortunately forces a fixed bound on base's
major version. But this conflicts with 1, and we judge this to be the
more important of the two (upgrades should be easy).